ml_services/requirements.txt
Flask==3.1.2
flask_cors==6.0.1
numpy==2.3.5
opencv_python==4.12.0.88
Pillow==12.0.0
torch==2.9.1
torchvision==0.24.1
Werkzeug==3.1.4
